+++ CHANGELOG.md
@@ -11,6 +11,7 @@ adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ### Changed
 - keepSuggestionsOnSuggestionSelected renamed to keepSuggestionsOnSelect
+- hideSuggestionsOnKeyboardHide renamed to hideOnUnfocus
 
 ### Fixed
 - enableSuggestions was not passed to TextField